@@ -29,7 +29,7 @@ class EggResource extends Resource
 
     public static function getNavigationGroup(): ?string
     {
-        return config('panel.filament.top-navigation', false) ? null : trans('admin/dashboard.server');
+        return !empty(auth()->user()->getCustomization()['top_navigation']) ? false : trans('admin/dashboard.server');
     }
 
     public static function getNavigationLabel(): string
