@@ -43,7 +43,7 @@ class ServerResource extends Resource
 
     public static function getNavigationGroup(): ?string
     {
-        return config('panel.filament.top-navigation', false) ? null : trans('admin/dashboard.server');
+        return !empty(auth()->user()->getCustomization()['top_navigation']) ? false : trans('admin/dashboard.server');
     }
 
     public static function getNavigationBadge(): ?string
