@@ -162,6 +162,7 @@ class Console extends Page
 
         return [
             Action::make('start')
+                ->label(trans('server/console.power_actions.start'))
                 ->color('primary')
                 ->size(ActionSize::ExtraLarge)
                 ->dispatch('setServerState', ['state' => 'start', 'uuid' => $server->uuid])
@@ -169,6 +170,7 @@ class Console extends Page
                 ->disabled(fn () => $server->isInConflictState() || !$this->status->isStartable())
                 ->icon('tabler-player-play-filled'),
             Action::make('restart')
+                ->label(trans('server/console.power_actions.restart'))
                 ->color('gray')
                 ->size(ActionSize::ExtraLarge)
                 ->dispatch('setServerState', ['state' => 'restart', 'uuid' => $server->uuid])
@@ -176,6 +178,7 @@ class Console extends Page
                 ->disabled(fn () => $server->isInConflictState() || !$this->status->isRestartable())
                 ->icon('tabler-reload'),
             Action::make('stop')
+                ->label(trans('server/console.power_actions.stop'))
                 ->color('danger')
                 ->size(ActionSize::ExtraLarge)
                 ->dispatch('setServerState', ['state' => 'stop', 'uuid' => $server->uuid])
@@ -184,13 +187,24 @@ class Console extends Page
                 ->disabled(fn () => $server->isInConflictState() || !$this->status->isStoppable())
                 ->icon('tabler-player-stop-filled'),
             Action::make('kill')
+                ->label(trans('server/console.power_actions.kill'))
                 ->color('danger')
-                ->tooltip('This can result in data corruption and/or data loss!')
+                ->tooltip(trans('server/console.power_actions.kill_tooltip'))
                 ->size(ActionSize::ExtraLarge)
                 ->dispatch('setServerState', ['state' => 'kill', 'uuid' => $server->uuid])
                 ->authorize(fn () => auth()->user()->can(Permission::ACTION_CONTROL_STOP, $server))
                 ->hidden(fn () => $server->isInConflictState() || !$this->status->isKillable())
                 ->icon('tabler-alert-square'),
         ];
+    }
+
+    public static function getNavigationLabel(): string
+    {
+        return trans('server/console.title');
+    }
+
+    public function getTitle(): string
+    {
+        return trans('server/console.title');
     }
 }
