@@ -2,37 +2,37 @@
 
 namespace App\Http\Controllers\Auth;
 
+use App\Extensions\OAuth\OAuthSchemaInterface;
 use App\Extensions\OAuth\OAuthService;
 use App\Filament\Pages\Auth\EditProfile;
 use App\Http\Controllers\Controller;
 use App\Models\User;
-use App\Services\Users\UserUpdateService;
+use App\Services\Users\UserCreationService;
 use Exception;
 use Filament\Notifications\Notification;
-use Illuminate\Auth\AuthManager;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Laravel\Socialite\Contracts\User as OAuthUser;
 use Laravel\Socialite\Facades\Socialite;
+use Symfony\Component\HttpFoundation\RedirectResponse as SymfonyRedirectResponse;
 
 class OAuthController extends Controller
 {
     public function __construct(
-        private readonly AuthManager $auth,
-        private readonly UserUpdateService $updateService,
-        private readonly OAuthService $oauthService
+        private readonly UserCreationService $userCreation,
+        private readonly OAuthService $oauthService,
     ) {}
 
     /**
      * Redirect user to the OAuth provider
      */
-    public function redirect(string $driver): RedirectResponse
+    public function redirect(string $driver): SymfonyRedirectResponse|RedirectResponse
     {
-        // Driver is disabled - redirect to normal login
         if (!$this->oauthService->get($driver)->isEnabled()) {
             return redirect()->route('auth.login');
         }
 
-        return Socialite::with($driver)->redirect();
+        return Socialite::driver($driver)->redirect();
     }
 
     /**
@@ -40,8 +40,9 @@ class OAuthController extends Controller
      */
     public function callback(Request $request, string $driver): RedirectResponse
     {
-        // Driver is disabled - redirect to normal login
-        if (!$this->oauthService->get($driver)?->isEnabled()) {
+        $driver = $this->oauthService->get($driver);
+
+        if (!$driver || !$driver->isEnabled()) {
             return redirect()->route('auth.login');
         }
 
@@ -49,43 +50,89 @@ class OAuthController extends Controller
         if ($request->get('error')) {
             report($request->get('error_description') ?? $request->get('error'));
 
-            Notification::make()
-                ->title('Something went wrong')
-                ->body($request->get('error'))
-                ->danger()
-                ->persistent()
-                ->send();
-
-            return redirect()->route('auth.login');
+            return $this->errorRedirect($request->get('error'));
         }
 
-        $oauthUser = Socialite::driver($driver)->user();
+        $oauthUser = Socialite::driver($driver->getId())->user();
 
-        // User is already logged in and wants to link a new OAuth Provider
         if ($request->user()) {
-            $oauth = $request->user()->oauth;
-            $oauth[$driver] = $oauthUser->getId();
-
-            $this->updateService->handle($request->user(), ['oauth' => $oauth]);
+            $this->linkUser($request->user(), $driver, $oauthUser);
 
             return redirect(EditProfile::getUrl(['tab' => '-oauth-tab'], panel: 'app'));
         }
 
-        try {
-            $user = User::query()->whereJsonContains('oauth->'. $driver, $oauthUser->getId())->firstOrFail();
-
-            $this->auth->guard()->login($user, true);
-        } catch (Exception) {
-            // No user found - redirect to normal login
-            Notification::make()
-                ->title('No linked User found')
-                ->danger()
-                ->persistent()
-                ->send();
-
-            return redirect()->route('auth.login');
+        $user = User::whereJsonContains('oauth->'. $driver->getId(), $oauthUser->getId())->first();
+        if ($user) {
+            return $this->loginUser($user);
         }
 
+        return $this->handleMissingUser($driver, $oauthUser);
+    }
+
+    private function linkUser(User $user, OAuthSchemaInterface $driver, OAuthUser $oauthUser): User
+    {
+        $oauth = $user->oauth;
+        $oauth[$driver->getId()] = $oauthUser->getId();
+
+        $user->update(['oauth' => $oauth]);
+
+        return $user->refresh();
+    }
+
+    private function handleMissingUser(OAuthSchemaInterface $driver, OAuthUser $oauthUser): RedirectResponse
+    {
+        $email = $oauthUser->getEmail();
+
+        if (!$email) {
+            return $this->errorRedirect();
+        }
+
+        $user = User::whereEmail($email)->first();
+        if ($user) {
+            if (!$driver->shouldLinkMissingUsers()) {
+                return $this->errorRedirect();
+            }
+
+            $user = $this->linkUser($user, $driver, $oauthUser);
+        } else {
+            if (!$driver->shouldCreateMissingUsers()) {
+                return $this->errorRedirect();
+            }
+
+            try {
+                $user = $this->userCreation->handle([
+                    'username' => $oauthUser->getNickname(),
+                    'email' => $email,
+                    'oauth' => [
+                        $driver->getId() => $oauthUser->getId(),
+                    ],
+                ]);
+            } catch (Exception $exception) {
+                report($exception);
+
+                return $this->errorRedirect();
+            }
+        }
+
+        return $this->loginUser($user);
+    }
+
+    private function loginUser(User $user): RedirectResponse
+    {
+        auth()->guard()->login($user, true);
+
         return redirect('/');
+    }
+
+    private function errorRedirect(?string $error = null): RedirectResponse
+    {
+        Notification::make()
+            ->title($error ? 'Something went wrong' : 'No linked User found')
+            ->body($error)
+            ->danger()
+            ->persistent()
+            ->send();
+
+        return redirect()->route('auth.login');
     }
 }
