@@ -4,25 +4,32 @@ namespace App\Livewire;
 
 use Closure;
 use Filament\Notifications\Concerns;
-use Filament\Support\Concerns\EvaluatesClosures;
-use Illuminate\Support\Str;
-use Livewire\Wireable;
+use Filament\Support\Components\ViewComponent;
+use Illuminate\Contracts\Support\Arrayable;
 
-final class AlertBanner implements Wireable
+final class AlertBanner extends ViewComponent implements Arrayable
 {
     use Concerns\HasBody;
     use Concerns\HasIcon;
     use Concerns\HasId;
     use Concerns\HasStatus;
     use Concerns\HasTitle;
-    use EvaluatesClosures;
 
     protected bool|Closure $closable = false;
 
-    public static function make(?string $id = null): AlertBanner
+    protected string $view = 'livewire.alerts.alert-banner';
+
+    protected string $viewIdentifier = 'alert-banner';
+
+    public function __construct(string $id)
     {
-        $static = new self();
-        $static->id($id ?? Str::orderedUuid());
+        $this->id($id);
+    }
+
+    public static function make(string $id): AlertBanner
+    {
+        $static = new self($id);
+        $static->configure();
 
         return $static;
     }
@@ -30,7 +37,7 @@ final class AlertBanner implements Wireable
     /**
      * @return array{id: string, title: ?string, body: ?string, status: ?string, icon: ?string, closeable: bool}
      */
-    public function toLivewire(): array
+    public function toArray(): array
     {
         return [
             'id' => $this->getId(),
@@ -42,15 +49,18 @@ final class AlertBanner implements Wireable
         ];
     }
 
-    public static function fromLivewire(mixed $value): AlertBanner
+    /**
+     * @param  array{id: string, title: ?string, body: ?string, status: ?string, icon: ?string, closeable: bool}  $data
+     */
+    public static function fromArray(array $data): AlertBanner
     {
-        $static = AlertBanner::make($value['id']);
+        $static = AlertBanner::make($data['id']);
 
-        $static->title($value['title']);
-        $static->body($value['body']);
-        $static->status($value['status']);
-        $static->icon($value['icon']);
-        $static->closable($value['closeable']);
+        $static->title($data['title']);
+        $static->body($data['body']);
+        $static->status($data['status']);
+        $static->icon($data['icon']);
+        $static->closable($data['closeable']);
 
         return $static;
     }
@@ -69,7 +79,7 @@ final class AlertBanner implements Wireable
 
     public function send(): AlertBanner
     {
-        session()->push('alert-banners', $this->toLivewire());
+        session()->push('alert-banners', $this->toArray());
 
         return $this;
     }
