@@ -24,37 +24,40 @@ class ServerEntry extends Component
                 style="background-color: #D97706;">
             </div>
 
-            <div class="flex-1 dark:bg-gray-850 dark:text-white rounded-lg overflow-hidden p-2">
+            <div class="flex-1 dark:bg-gray-800 dark:text-white rounded-lg overflow-hidden p-3">
                 <div class="flex items-center mb-5 gap-2">
-                    <x-filament::loading-indicator class="h-5 w-5" />
+                    <x-filament::loading-indicator class="h-6 w-6" />
                     <h2 class="text-xl font-bold">
                         {{ $server->name }}
+                        <span class="dark:text-gray-400">
+                        ({{ trans('server/dashboard.loading') }})
+                        </span>
                     </h2>
                 </div>
 
-                <div class="flex justify-between text-center">
+                <div class="flex justify-between text-center items-center gap-4">
                     <div>
-                        <p class="text-sm dark:text-gray-400">CPU</p>
+                        <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.cpu') }}</p>
                         <p class="text-md font-semibold">{{ Number::format(0, precision: 2, locale: auth()->user()->language ?? 'en') . '%' }}</p>
                         <hr class="p-0.5">
-                        <p class="text-xs dark:text-gray-400">{{ $server->formatResource('cpu', type: \App\Enums\ServerResourceType::Percentage, limit: true) }}</p>
+                        <p class="text-xs dark:text-gray-400">{{ $server->formatResource(\App\Enums\ServerResourceType::CPULimit) }}</p>
                     </div>
                     <div>
-                        <p class="text-sm dark:text-gray-400">Memory</p>
+                        <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.memory') }}</p>
                         <p class="text-md font-semibold">{{ convert_bytes_to_readable(0, decimals: 2) }}</p>
                         <hr class="p-0.5">
-                        <p class="text-xs dark:text-gray-400">{{ $server->formatResource('memory', limit: true) }}</p>
+                        <p class="text-xs dark:text-gray-400">{{ $server->formatResource(\App\Enums\ServerResourceType::MemoryLimit) }}</p>
                     </div>
                     <div>
-                        <p class="text-sm dark:text-gray-400">Disk</p>
+                        <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.disk') }}</p>
                         <p class="text-md font-semibold">{{ convert_bytes_to_readable(0, decimals: 2) }}</p>
                         <hr class="p-0.5">
-                        <p class="text-xs dark:text-gray-400">{{ $server->formatResource('disk', limit: true) }}</p>
+                        <p class="text-xs dark:text-gray-400">{{ $server->formatResource(\App\Enums\ServerResourceType::DiskLimit) }}</p>
                     </div>
                     <div class="hidden sm:block">
-                        <p class="text-sm dark:text-gray-400">Network</p>
+                        <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.network') }}</p>
                         <hr class="p-0.5">
-                        <p class="text-md font-semibold">{{ $server->allocation?->address ?? 'None' }} </p>
+                        <p class="text-md font-semibold">{{ $server->allocation?->address ?? trans('server/dashboard.none') }} </p>
                     </div>
                 </div>
             </div>
