@@ -2,68 +2,27 @@
 
 namespace App\Providers\Filament;
 
-use App\Filament\Pages\Auth\Login;
-use App\Filament\Pages\Auth\EditProfile;
-use App\Http\Middleware\LanguageMiddleware;
-use App\Http\Middleware\RequireTwoFactorAuthentication;
 use Filament\Facades\Filament;
-use Filament\Http\Middleware\Authenticate;
-use Filament\Http\Middleware\DisableBladeIconComponents;
-use Filament\Http\Middleware\DispatchServingFilamentEvent;
 use Filament\Navigation\MenuItem;
 use Filament\Panel;
-use Filament\PanelProvider;
-use Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse;
-use Illuminate\Cookie\Middleware\EncryptCookies;
-use Illuminate\Foundation\Http\Middleware\VerifyCsrfToken;
-use Illuminate\Routing\Middleware\SubstituteBindings;
-use Illuminate\Session\Middleware\AuthenticateSession;
-use Illuminate\Session\Middleware\StartSession;
-use Illuminate\View\Middleware\ShareErrorsFromSession;
 
 class AppPanelProvider extends PanelProvider
 {
     public function panel(Panel $panel): Panel
     {
-        return $panel
+        return parent::panel($panel)
             ->id('app')
-            ->spa()
-            ->databaseNotifications()
+            ->default()
             ->breadcrumbs(false)
-            ->brandName(config('app.name', 'Pelican'))
-            ->brandLogo(config('app.logo'))
-            ->brandLogoHeight('2rem')
-            ->favicon(config('app.favicon', '/pelican.ico'))
-            ->topNavigation(config('panel.filament.top-navigation', false))
-            ->maxContentWidth(config('panel.filament.display-width', 'screen-2xl'))
             ->navigation(false)
-            ->profile(EditProfile::class, false)
-            ->login(Login::class)
-            ->passwordReset()
             ->userMenuItems([
                 MenuItem::make()
-                    ->label('Admin')
+                    ->label(trans('profile.admin'))
                     ->url('/admin')
                     ->icon('tabler-arrow-forward')
                     ->sort(5)
-                    ->visible(fn (): bool => auth()->user()->canAccessPanel(Filament::getPanel('admin'))),
+                    ->visible(fn () => auth()->user()->canAccessPanel(Filament::getPanel('admin'))),
             ])
-            ->discoverResources(in: app_path('Filament/App/Resources'), for: 'App\\Filament\\App\\Resources')
-            ->middleware([
-                EncryptCookies::class,
-                AddQueuedCookiesToResponse::class,
-                StartSession::class,
-                AuthenticateSession::class,
-                ShareErrorsFromSession::class,
-                VerifyCsrfToken::class,
-                SubstituteBindings::class,
-                DisableBladeIconComponents::class,
-                DispatchServingFilamentEvent::class,
-                LanguageMiddleware::class,
-                RequireTwoFactorAuthentication::class,
-            ])
-            ->authMiddleware([
-                Authenticate::class,
-            ]);
+            ->discoverResources(in: app_path('Filament/App/Resources'), for: 'App\\Filament\\App\\Resources');
     }
 }
