@@ -20,7 +20,14 @@ class EggConfigurationService
      * @return array{
      *     startup: array{done: string[], user_interaction: string[], strip_ansi: bool},
      *     stop: array{type: string, value: string},
-     *     configs: array<mixed>
+     *     configs: list<array{
+     *         file: string,
+     *         replace: list<array{
+     *             match: string,
+     *             if_value?: string,
+     *             replace_with: string
+     *         }>
+     *     }>
      * }
      */
     public function handle(Server $server): array
@@ -81,9 +88,10 @@ class EggConfigurationService
     }
 
     /**
-     * @return array<mixed>
+     * @param  array<string, mixed>  $configs
+     * @return array<int, array<string, mixed>>
      */
-    protected function replacePlaceholders(Server $server, object $configs): array
+    protected function replacePlaceholders(Server $server, object|array $configs): array
     {
         // Get the legacy configuration structure for the server so that we
         // can property map the egg placeholders to values.
