@@ -3,6 +3,7 @@
 namespace App\Services\Eggs\Sharing;
 
 use App\Exceptions\Service\InvalidFileUploadException;
+use JsonException;
 use Ramsey\Uuid\Uuid;
 use Illuminate\Support\Arr;
 use App\Models\Egg;
@@ -11,6 +12,9 @@ use App\Models\EggVariable;
 use Illuminate\Database\ConnectionInterface;
 use Illuminate\Support\Collection;
 use Spatie\TemporaryDirectory\TemporaryDirectory;
+use stdClass;
+use Symfony\Component\Yaml\Yaml;
+use Throwable;
 
 class EggImporterService
 {
@@ -28,9 +32,9 @@ class EggImporterService
     public function __construct(protected ConnectionInterface $connection) {}
 
     /**
-     * Take an uploaded JSON file and parse it into a new egg.
+     * Take an uploaded JSON or YAML file and parse it into a new egg.
      *
-     * @throws \App\Exceptions\Service\InvalidFileUploadException|\Throwable
+     * @throws InvalidFileUploadException|Throwable
      */
     public function fromFile(UploadedFile $file, ?Egg $egg = null): Egg
     {
@@ -46,7 +50,6 @@ class EggImporterService
                 'copy_script_from' => null,
             ]);
 
-            // Don't check for this anymore
             for ($i = 0; $i < count($parsed['variables']); $i++) {
                 unset($parsed['variables'][$i]['field_type']);
             }
@@ -54,7 +57,6 @@ class EggImporterService
             $egg = $this->fillFromParsed($egg, $parsed);
             $egg->save();
 
-            // Update existing variables or create new ones.
             foreach ($parsed['variables'] ?? [] as $variable) {
                 EggVariable::unguarded(function () use ($egg, $variable) {
                     $variable['rules'] = is_array($variable['rules']) ? $variable['rules'] : explode('|', $variable['rules']);
@@ -66,7 +68,6 @@ class EggImporterService
             }
 
             $imported = array_map(fn ($value) => $value['env_variable'], $parsed['variables'] ?? []);
-
             $egg->variables()->whereNotIn('env_variable', $imported)->delete();
 
             return $egg->refresh();
@@ -74,31 +75,39 @@ class EggImporterService
     }
 
     /**
-     * Take an url and parse it into a new egg or update an existing one.
+     * Take a URL (YAML or JSON) and parse it into a new egg or update an existing one.
      *
-     * @throws \App\Exceptions\Service\InvalidFileUploadException|\Throwable
+     * @throws InvalidFileUploadException|Throwable
      */
     public function fromUrl(string $url, ?Egg $egg = null): Egg
     {
         $info = pathinfo($url);
+        $extension = strtolower($info['extension']);
+
         $tmpDir = TemporaryDirectory::make()->deleteWhenDestroyed();
         $tmpPath = $tmpDir->path($info['basename']);
 
-        if (!file_put_contents($tmpPath, file_get_contents($url))) {
-            throw new InvalidFileUploadException('Could not write temporary file.');
+        $fileContents = @file_get_contents($url);
+
+        if (!$fileContents || !file_put_contents($tmpPath, $fileContents)) {
+            throw new InvalidFileUploadException('Could not download or write temporary file.');
         }
 
-        return $this->fromFile(new UploadedFile($tmpPath, $info['basename'], 'application/json'), $egg);
+        $mime = match ($extension) {
+            'yaml', 'yml' => 'application/yaml',
+            'json' => 'application/json',
+            default => throw new InvalidFileUploadException('Unsupported file format.'),
+        };
+
+        return $this->fromFile(new UploadedFile($tmpPath, $info['basename'], $mime), $egg);
     }
 
     /**
      * Takes an uploaded file and parses out the egg configuration from within.
      *
-     * @todo replace with DTO
-     *
      * @return array<array-key, mixed>
      *
-     * @throws \App\Exceptions\Service\InvalidFileUploadException
+     * @throws InvalidFileUploadException|JsonException
      */
     protected function parseFile(UploadedFile $file): array
     {
@@ -106,30 +115,56 @@ class EggImporterService
             throw new InvalidFileUploadException('The selected file was not uploaded successfully');
         }
 
+        $extension = strtolower($file->getClientOriginalExtension());
+        $mime = $file->getMimeType();
+
         try {
-            $parsed = json_decode($file->getContent(), true, 512, JSON_THROW_ON_ERROR);
-        } catch (\JsonException $exception) {
-            throw new InvalidFileUploadException('Could not read JSON file: ' . $exception->getMessage());
+            $content = $file->getContent();
+
+            $parsed = match (true) {
+                in_array($extension, ['yaml', 'yml']),
+                str_contains($mime, 'yaml') => Yaml::parse($content),
+                default => json_decode($content, true, 512, JSON_THROW_ON_ERROR),
+            };
+        } catch (Throwable $e) {
+            throw new InvalidFileUploadException('File parse failed: ' . $e->getMessage());
         }
 
         $version = $parsed['meta']['version'] ?? '';
 
         $parsed = match ($version) {
             'PTDL_v1' => $this->convertToV2($parsed),
-            'PTDL_v2' => $parsed,
-            'PLCN_v1' => $parsed,
-            default => throw new InvalidFileUploadException('The JSON file provided is not in a format that can be recognized.')
+            'PTDL_v2', 'PLCN_v1', 'PLCN_v2' => $parsed,
+            default => throw new InvalidFileUploadException('The file format is not recognized.'),
         };
 
-        // Make sure we only use recent variable format from now on
-        if (array_get($parsed['config'], 'files')) {
-            $parsed['config']['files'] = str_replace(
-                array_keys(self::UPGRADE_VARIABLES),
-                array_values(self::UPGRADE_VARIABLES),
-                $parsed['config']['files'],
-            );
+        if (isset($parsed['config']) && (is_array($parsed['config']) || $parsed['config'] instanceof stdClass)) {
+            $parsed['config'] = (array) $parsed['config'];
+            foreach ($parsed['config'] as $key => $value) {
+                if (is_array($value) || $value instanceof stdClass) {
+                    $parsed['config'][$key] = json_encode((array) $value, JSON_UNESCAPED_SLASHES | JSON_THROW_ON_ERROR);
+                }
+
+                if ($key === 'files' && is_string($parsed['config'][$key])) {
+                    $parsed['config'][$key] = str_replace(
+                        array_keys(self::UPGRADE_VARIABLES),
+                        array_values(self::UPGRADE_VARIABLES),
+                        $parsed['config'][$key]
+                    );
+                }
+            }
         }
 
+        if (isset($parsed['scripts']['installation']) && (is_array($parsed['scripts']['installation']) || $parsed['scripts']['installation'] instanceof stdClass)) {
+            $parsed['scripts']['installation'] = (array) $parsed['scripts']['installation'];
+            foreach ($parsed['scripts']['installation'] as $key => $value) {
+                if (is_array($value) || $value instanceof stdClass) {
+                    $parsed['scripts']['installation'][$key] = json_encode((array) $value, JSON_UNESCAPED_SLASHES | JSON_THROW_ON_ERROR);
+                }
+            }
+        }
+
+        // Reserved env var name handling
         [$forbidden, $allowed] = collect($parsed['variables'])
             ->map(fn ($variable) => array_merge(
                 $variable,
@@ -155,22 +190,7 @@ class EggImporterService
     }
 
     /**
-     * Fills the provided model with the parsed JSON data.
-     *
-     * @param array{
-     *     name: string,
-     *     description: string,
-     *     tags: string[],
-     *     features: string[],
-     *     docker_images: string[],
-     *     file_denylist: string[],
-     *     meta: array{update_url: string},
-     *     config: array{files: string, startup: string, logs: string, stop: string},
-     *     startup: string,
-     *     scripts: array{
-     *         installation: array{script: string, entrypoint: string, container: string},
-     *     },
-     * } $parsed
+     * @param  array<string, mixed>  $parsed
      */
     protected function fillFromParsed(Egg $model, array $parsed): Egg
     {
@@ -182,9 +202,9 @@ class EggImporterService
             'docker_images' => Arr::get($parsed, 'docker_images'),
             'file_denylist' => Collection::make(Arr::get($parsed, 'file_denylist'))->filter(fn ($value) => !empty($value)),
             'update_url' => Arr::get($parsed, 'meta.update_url'),
-            'config_files' => Arr::get($parsed, 'config.files'),
-            'config_startup' => Arr::get($parsed, 'config.startup'),
-            'config_logs' => Arr::get($parsed, 'config.logs'),
+            'config_files' => json_encode(json_decode(Arr::get($parsed, 'config.files')), JSON_PRETTY_PRINT),
+            'config_startup' => json_encode(json_decode(Arr::get($parsed, 'config.startup')), JSON_PRETTY_PRINT),
+            'config_logs' => json_encode(json_decode(Arr::get($parsed, 'config.logs')), JSON_PRETTY_PRINT),
             'config_stop' => Arr::get($parsed, 'config.stop'),
             'startup' => Arr::get($parsed, 'startup'),
             'script_install' => Arr::get($parsed, 'scripts.installation.script'),
@@ -194,17 +214,11 @@ class EggImporterService
     }
 
     /**
-     * Converts a PTDL_V1 egg into the expected PTDL_V2 egg format. This just handles
-     * the "docker_images" field potentially not being present, and not being in the
-     * expected "key => value" format.
-     *
-     * @param  array{images?: string[], image?: string, field_type?: string, docker_images?: array<array-key, string>}  $parsed
-     * @return array<array-key, mixed>
+     * @param  array<string, mixed>  $parsed
+     * @return array<string, mixed>
      */
     protected function convertToV2(array $parsed): array
     {
-        // Maintain backwards compatability for eggs that are still using the old single image
-        // string format. New eggs can provide an array of Docker images that can be used.
         if (!isset($parsed['images'])) {
             $images = [Arr::get($parsed, 'image') ?? 'nil'];
         } else {
