@@ -39,15 +39,7 @@ class ServerCreationService
      * as possible given the input data. For example, if an allocation_id is passed with
      * no node_id the node_is will be picked from the allocation.
      *
-     * @param array{
-     *     node_id?: int,
-     *     oom_killer?: bool,
-     *     oom_disabled?: bool,
-     *     egg_id?: int,
-     *     image?: ?string,
-     *     startup?: ?string,
-     *     start_on_completion?: ?bool,
-     * } $data
+     * @param  array<mixed, mixed>  $data
      *
      * @throws \Throwable
      * @throws \App\Exceptions\DisplayException
@@ -64,8 +56,8 @@ class ServerCreationService
         $egg = Egg::query()->findOrFail($data['egg_id']);
 
         // Fill missing fields from egg
-        $data['image'] = $data['image'] ?? collect($egg->docker_images)->first();
-        $data['startup'] = $data['startup'] ?? $egg->startup;
+        $data['image'] ??= collect($egg->docker_images)->first();
+        $data['startup'] ??= $egg->startup;
 
         // If a deployment object has been passed we need to get the allocation and node that the server should use.
         if ($deployment) {
@@ -94,6 +86,8 @@ class ServerCreationService
             if (empty($data['node_id'])) {
                 $data['node_id'] = $nodes->first();
             }
+        } else {
+            $data['node_id'] ??= Allocation::find($data['allocation_id'])?->node_id;
         }
 
         Assert::false(empty($data['node_id']), 'Expected a non-empty node_id in server creation data.');
