@@ -10,7 +10,7 @@ use App\Services\Nodes\NodeJWTService;
 use Carbon\CarbonImmutable;
 use Illuminate\Database\ConnectionInterface;
 use Illuminate\Support\Facades\Http;
-use Lcobucci\JWT\Token\Plain;
+use Lcobucci\JWT\UnencryptedToken;
 
 class TransferServerService
 {
@@ -22,7 +22,7 @@ class TransferServerService
         private NodeJWTService $nodeJWTService,
     ) {}
 
-    private function notify(ServerTransfer $transfer, Plain $token): void
+    private function notify(ServerTransfer $transfer, UnencryptedToken $token): void
     {
         Http::daemon($transfer->oldNode)->post("/api/servers/{$transfer->server->uuid}/transfer", [
             'url' => $transfer->newNode->getConnectionAddress() . '/api/transfers',
