@@ -3,10 +3,12 @@
 namespace Database\Seeders;
 
 use App\Models\Egg;
-use Exception;
+use DirectoryIterator;
 use Illuminate\Database\Seeder;
 use Illuminate\Http\UploadedFile;
 use App\Services\Eggs\Sharing\EggImporterService;
+use Symfony\Component\Yaml\Yaml;
+use Throwable;
 
 class EggSeeder extends Seeder
 {
@@ -46,22 +48,39 @@ class EggSeeder extends Seeder
      */
     protected function parseEggFiles($name): void
     {
-        $files = new \DirectoryIterator(database_path('Seeders/eggs/' . kebab_case($name)));
+        $path = database_path('Seeders/eggs/' . kebab_case($name));
+        $files = new DirectoryIterator($path);
 
         $this->command->alert('Updating Eggs for: ' . $name);
-        /** @var \DirectoryIterator $file */
+
+        /** @var DirectoryIterator $file */
         foreach ($files as $file) {
             if (!$file->isFile() || !$file->isReadable()) {
                 continue;
             }
 
+            $extension = strtolower($file->getExtension());
+            $filePath = $file->getRealPath();
+
             try {
-                $decoded = json_decode(file_get_contents($file->getRealPath()), true, 512, JSON_THROW_ON_ERROR);
-            } catch (Exception) {
+                $decoded = match ($extension) {
+                    'json' => json_decode(file_get_contents($filePath), true, 512, JSON_THROW_ON_ERROR),
+                    'yaml', 'yml' => Yaml::parseFile($filePath),
+                    default => null,
+                };
+            } catch (Throwable) {
+                $this->command->warn("Failed to parse {$file->getFilename()}, skipping.");
+
                 continue;
             }
 
-            $file = new UploadedFile($file->getPathname(), $file->getFilename(), 'application/json');
+            if (!is_array($decoded) || !isset($decoded['name'], $decoded['author'])) {
+                $this->command->warn("Invalid structure in {$file->getFilename()}, skipping.");
+
+                continue;
+            }
+
+            $uploaded = new UploadedFile($filePath, $file->getFilename());
 
             $egg = Egg::query()
                 ->where('author', $decoded['author'])
@@ -69,10 +88,10 @@ class EggSeeder extends Seeder
                 ->first();
 
             if ($egg instanceof Egg) {
-                $this->importerService->fromFile($file, $egg);
+                $this->importerService->fromFile($uploaded, $egg);
                 $this->command->info('Updated ' . $decoded['name']);
             } else {
-                $this->importerService->fromFile($file);
+                $this->importerService->fromFile($uploaded);
                 $this->command->comment('Created ' . $decoded['name']);
             }
         }
