@@ -13,13 +13,17 @@ return [
     'import' => [
         'file' => 'File',
         'url' => 'URL',
-        'egg_help' => 'This should be the raw .json file ( egg-minecraft.json )',
-        'url_help' => 'URLs must point directly to the raw .json file',
+        'egg_help' => 'This should be the raw .json/.yaml file',
+        'url_help' => 'URLs must point directly to the raw .json/.yaml file',
         'add_url' => 'New URL',
         'import_failed' => 'Import Failed',
         'import_success' => 'Import Success',
         'github' => 'Add from Github',
         'refresh' => 'Refresh',
+    ],
+    'export' => [
+        'modal' => 'How would you like to export :egg ?',
+        'as' => 'As',
     ],
     'in_use' => 'In Use',
     'servers' => 'Servers',
