@@ -12,6 +12,7 @@ return [
         'customization' => 'Customization',
     ],
     'username' => 'Username',
+    'admin' => 'Admin',
     'exit_admin' => 'Exit Admin',
     'email' => 'Email',
     'password' => 'Password',
@@ -20,8 +21,8 @@ return [
     'timezone' => 'Timezone',
     'language' => 'Language',
     'language_help' => 'Your language :state has not been translated yet!',
-    'link' => 'Link ',
-    'unlink' => 'Unlink ',
+    'link' => 'Link :name',
+    'unlink' => 'Unlink :name',
     'unlinked' => ':name unlinked',
     'scan_qr' => 'Scan QR Code',
     'code' => 'Code',
@@ -57,4 +58,7 @@ return [
     'seconds' => 'Seconds',
     'graph_period' => 'Graph Period',
     'graph_period_helper' => 'The amount of data points, seconds, shown on the console graphs.',
+    'navigation' => 'Navigation Type',
+    'top' => 'Topbar',
+    'side' => 'Sidebar',
 ];
