@@ -30,8 +30,8 @@
                 class="w-full focus:outline-none focus:ring-0 border-none dark:bg-gray-900"
                 type="text"
                 :readonly="{{ $this->canSendCommand() ? 'false' : 'true' }}"
-                title="{{ $this->canSendCommand() ? '' : 'Can\'t send command when the server is Offline' }}"
-                placeholder="{{ $this->canSendCommand() ? 'Type a command...' : 'Server Offline...' }}"
+                title="{{ $this->canSendCommand() ? '' : trans('server/console.command_blocked_title') }}"
+                placeholder="{{ $this->canSendCommand() ? trans('server/console.command') : trans('server/console.command_blocked') }}"
                 wire:model="input"
                 wire:keydown.enter="enter"
                 wire:keydown.up.prevent="up"
