@@ -1,5 +1,5 @@
 <div id="alert-banner-container" class="flex flex-col gap-4">
-    @foreach (array_values($alertBanners) as $alertBanner)
-        @include('livewire.alerts.alert-banner', ['alertBanner' => $alertBanner])
+    @foreach ($alertBanners as $alertBanner)
+        {{ $alertBanner }}
     @endforeach
 </div>
