@@ -1,14 +1,12 @@
-@props(['alertBanner'])
-
 @php
-    $icon = $alertBanner->getIcon();
-    $title = $alertBanner->getTitle();
-    $body = $alertBanner->getBody();
+    $icon = $getIcon();
+    $title = $getTitle();
+    $body = $getBody();
 @endphp
 
-<div class="{{$alertBanner->getColorClasses()}} flex p-4 mt-3 rounded-xl shadow-lg bg-white dark:bg-gray-900 ring-1 ring-gray-950/5 dark:ring-white/10">
+<div class="{{$getColorClasses()}} flex p-4 mt-3 rounded-xl shadow-lg bg-white dark:bg-gray-900 ring-1 ring-gray-950/5 dark:ring-white/10">
     @if (filled($icon))
-        <x-filament::icon :icon="$icon" class="h-8 w-8 mr-2" color="{{$alertBanner->getStatus()}}" />
+        <x-filament::icon :icon="$icon" class="h-8 w-8 mr-2" color="{{$getStatus()}}" />
     @endif
 
     <div class="flex flex-col flex-grow">
@@ -21,7 +19,7 @@
         @endif
     </div>
 
-    @if ($alertBanner->isCloseable())
-        <x-filament::icon-button color="gray" icon="tabler-x" wire:click="remove('{{$alertBanner->getID()}}')" />
+    @if ($isCloseable())
+        <x-filament::icon-button color="gray" icon="tabler-x" wire:click="remove('{{$getId()}}')" />
     @endif
 </div>
