@@ -17,7 +17,7 @@
             <h2 class="text-xl font-bold">
                 {{ $server->name }}
                 <span class="dark:text-gray-400">
-                    ({{ $server->formatResource('uptime', type: \App\Enums\ServerResourceType::Time) }})
+                    ({{ $server->formatResource(\App\Enums\ServerResourceType::Uptime) }})
                 </span>
             </h2>
             <div class="end-0" x-on:click.stop>
@@ -33,27 +33,27 @@
 
         <div class="flex justify-between text-center items-center gap-4">
             <div>
-                <p class="text-sm dark:text-gray-400">CPU</p>
-                <p class="text-md font-semibold">{{ $server->formatResource('cpu_absolute', type: \App\Enums\ServerResourceType::Percentage) }}</p>
+                <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.cpu') }}</p>
+                <p class="text-md font-semibold">{{ $server->formatResource(\App\Enums\ServerResourceType::CPU) }}</p>
                 <hr class="p-0.5">
-                <p class="text-xs dark:text-gray-400">{{ $server->formatResource('cpu', limit: true, type: \App\Enums\ServerResourceType::Percentage) }}</p>
+                <p class="text-xs dark:text-gray-400">{{ $server->formatResource(\App\Enums\ServerResourceType::CPULimit) }}</p>
             </div>
             <div>
-                <p class="text-sm dark:text-gray-400">Memory</p>
-                <p class="text-md font-semibold">{{ $server->formatResource('memory_bytes') }}</p>
+                <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.memory') }}</p>
+                <p class="text-md font-semibold">{{ $server->formatResource(\App\Enums\ServerResourceType::Memory) }}</p>
                 <hr class="p-0.5">
-                <p class="text-xs dark:text-gray-400">{{ $server->formatResource('memory', limit: true) }}</p>
+                <p class="text-xs dark:text-gray-400">{{ $server->formatResource(\App\Enums\ServerResourceType::MemoryLimit) }}</p>
             </div>
             <div>
-                <p class="text-sm dark:text-gray-400">Disk</p>
-                <p class="text-md font-semibold">{{ $server->formatResource('disk_bytes') }}</p>
+                <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.disk') }}</p>
+                <p class="text-md font-semibold">{{ $server->formatResource(\App\Enums\ServerResourceType::Disk) }}</p>
                 <hr class="p-0.5">
-                <p class="text-xs dark:text-gray-400">{{ $server->formatResource('disk', limit: true) }}</p>
+                <p class="text-xs dark:text-gray-400">{{ $server->formatResource(\App\Enums\ServerResourceType::DiskLimit) }}</p>
             </div>
             <div class="hidden sm:block">
-                <p class="text-sm dark:text-gray-400">Network</p>
+                <p class="text-sm dark:text-gray-400">{{ trans('server/dashboard.network') }}</p>
                 <hr class="p-0.5">
-                <p class="text-md font-semibold">{{ $server->allocation?->address ?? 'None' }}</p>
+                <p class="text-md font-semibold">{{ $server->allocation?->address ?? trans('server/dashboard.none') }}</p>
             </div>
         </div>
     </div>
