@@ -5,9 +5,11 @@ export default {
     presets: [preset],
     content: [
         './app/Filament/**/*.php',
-        './resources/views/filament/**/*.blade.php',
-        './vendor/filament/**/*.blade.php',
         './app/Livewire/**/*.php',
+
+        './resources/views/filament/**/*.blade.php',
         './resources/views/livewire/**/*.blade.php',
+
+        './vendor/filament/**/*.blade.php',
     ],
 };
